// schmeconomics.client/app/lib/openapi/model/createCategoryRequest.ts
/**
 * Schmeconomics.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateCategoryRequest { 
    accountId: string;
    name: string;
    balance: number;
    refillValue: number;
}